feat(avataredit): convert to typescript
